Squashed commit: use scx15 source
This is the entirety of patches that made it possible to run android marshmallow
on SM-G355H (not sure on other SM-G355** devices though).

Everything has been put as it is in the previous branch. Nothing lost or added.

Since ih24n69 did most of the work, ih24n69 deserves to be the author of this commit.

But seriously since Nougat, most of the initial patchset has been reduced to this.
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>