gralloc: scx15: Prepare simple mitigations to get HIDL working
This patch contains two workarounds.

The first workaround creates a dummy but otherwise working fd as a
replacement to the invalid fd. The dummy fd should allow both cloning
and transport (via hwbinder) to work properly.

The second workaround prevents the free() callback upon framebuffer
memory. This is to avoid the new way of allocating a buffer where
the free() callback is called inside allocate(). I don't know why it is
done but the framebuffer memory is never allocated in the first place
only mmap'd once then it's the same forever.

Change-Id: Ide79bbb09c666e096cd07bcc5c4ecd1d0d389614
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>