gralloc: scx15: Expose USE_3_FRAMEBUFFER flag
On your BoardConfig.mk, set USE_3_FRAMEBUFFER to true.

This makes the gralloc module interrogate the FB (framebufer) to use
a contiguous display buffer the size of three displays.

Usually this is not needed, but on Android 7 and above,
the failure to query for the display parameters not only
disables page-flipping but also borks the dpi parameters.
This will cause some apps that depend on screen dpi,
for scaling, to display ant size images or elements.

Change-Id: Ie060bb73c1776f5fecff3ab487094c97b11b8a5f
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>