hwcomposer:scx15: Use OverlayComposer from the sc8830 side
It turns out that the OverlayComposer on the sc8830 is more up-to-date
than what we have now. And, migrating over that version is as simple
as using the same code without alterations.

The main reason to use this instead is to solve the bug where
the ui elements would be stretched out of the left side of the screen
when a video layer is present (e.g. playing a video on
the YouTube app)

To test this:
1. Play a video in the YouTube app, or some local video file via
the AOSP's "Video Player" app.
2. Make sure the video is playing in portrait mode; this makes the
bug obvious.
3. Adjust the volume to show Android's native volume slider.
4. Check whether the volume slider "dialog" "underflows" on the left side
of the screen. It should not flow outside of the screen.

Change-Id: I706117b873280d154fb1e0e794663311e1bdfed3
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>