omx: avc: scx15: Allow Constant Bitrate input to the encoder.
This is not fully tested, but the encoder (both the library and the harware)
should be able to handle the constant bitrates just fine.
Actually, the library and the hardware decoder does not seem to care at all...

This patch should have a wonderful side-effect of making screen mirroring
or screen casting possible. At least, it's successful using Chromecast.

Change-Id: I46ff3be04e84f04c55367832566bf00b4f0aeed4
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>