omx: avcenc: scx15: Latency tradeoffs on AndroidOpaque inputs larger than 480x800x2
The encoder seriously has some problems when the resolution is way beyond
it's intended capacity.

P-frames is the default output to conserve bandwidth by since it only sends the
"deltas" between the frame but it requires additional work for the encoder.
Where as I-frames just takes the whole image in a frame which should save some time
but at a cost of requiring a *much* higher bandwidth
(about 1-4 MegaBytes of bandwidth which should be okay for wifi direct).

To alleviate the high bandwidth requirement, the quantization parameter has been
increased and rate control has been disabled. This will reduce the bitrate at the
cost of quality and details.

This should drastically speed-up screen mirror/casting on middle to low tire devices.

PS. the rate control is whack as it fails to find the "just right" parameter
and causes a pulsating effect on the encoded video where the QP is seen
gradually rising then sharply falling afterwards.

PS2. Screen recording with the an area less than twice of 480*800 will not trigger
this change. The tradeoff occurs iff the area is at least twice larger.

Change-Id: I9819a43bf73923afafecb734101d63d8a78d8318
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>