audio: scx15: fix parsing for <debuginfo> xml tag
After <voip> tag is parsed,
any tags defined in the else if chain
of conditions, after the condition for the voip tag, will not be parse
anymore.

In simpler terms, within the audio_hw.xml file, ifever a <debug>
or <debuginfo> tag is placed after the <voip> tag, it
will never be parse properly.

The fix is a simple refactor of the condition.

Change-Id: Idbf3ddf9504449c730e1c89c97d912fd17a1ff58
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>