kanas: Switch to sc8830 (previously scx30g) avc (h.264) decoder
The migration is straightforward enough. This requires 2 functions from
libstagefrieght_sprd and would have API problems with memoryheap_sprd
which explains the lots of macros working around this.

Kanas doesn't use IOMMU (or has one) so testing the IOMMU part is nada.
Though this commit is aimed for kanas, this should work for scx15 devices in general.
But, the binary blob (the actual h.264 encoder) needs one function to be
binary patched to invert the clock indices so that the
right clock is being set. That binary blob must be
for the scx30g version. I'm using the 2017 one.

Change-Id: I3b939faf7ec9cb667aab5510b347c8c4d117dfaf
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>