audio: scx15: Maximize Long Periods and minimize Short periods
By convention, normal playbacks (short periods) usually have a latency around 20ms
and deep buffers (long periods) have around 100ms. And, IRAM and the
regular buffer size should be in 1:4 ratio.

Unfortunately using IRAM, reaching 100 ms is impossible. Though, we could actually
maximize the size as we're not using it fully any way.

To make it fill all the way for the long period, we might want to reduce the short period
multiplier to 5 (from 8) so that we could get a divisible number for the long period.
This patch should entirely maximize the long period buffer size and theoretically
cannot be expanded any further (I still haven't found my scratch for the math)

The effect is that a less delayed sound in normal playback and a slightly larger
(yet inconceivable for IRAM only) deep buffer.

Based on a idle test with 76 sample size, using a 90ms deep buffer IRAM
against 360ms of the regular deep buffer. The one with IRAM is 0.5% more idle
than the other. And, that is pretty much worthless.

Change-Id: Ife82baa113003d9c6e9781010a07d64e09bc7f55
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>