libcamera: kanas: scx15: Enable s5k4ec mipi sensor, move HI702 to index 2
This was disabled due to 5066fc1885ca313bd868ab9fbc82410542618d78
(camera: scx15: fix compile for N)

But, kanas3g phone needs this. Will fix the s5k4ecgx mipi sensor
to feature parity later on.

TL;DR; When switching over to the stock or opensource impl of SPRD libcamera HAL
delete the "/data/misc/media/sensor.file"

And, to point it out, that specific change could make the cached
identification to be invalid if ever someone did used the stock libcamera
HAL before this opensource one. The HAL seems to never clear that cache so
that it will persist until the user manually removes the file at:
/data/misc/media/sensor.file

The guys at SPRD realized that doing an identify to all of that
other sensors (the ones that the change commented ou) will take some time
and it has to do it every time the HAL is init'd (or every time you use
either of the camera). So they did it once and saved the results to
a file, once.

This stale cache may cause the identifier to fail to load the driver
and may even fail to power it down when initiallizing the other sensor
which may also make that other sensor fail at init.

Change-Id: Ifdb85b867d2afdb697e094f3f52121b1b16c6aea
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>