libcamera: scx15: Implement kMetadataBufferTypeNativeHandleSource for HAL1.0
This is a drastically simplified version of djeman's commit in his vendor tree
https://github.com/djeman/android_vendor_sprd/commit/bd2c631df83c480d12a00a2d50a7881510e7fa40
which should implement kMetadataBufferTypeNativeHandleSource needed for Android N and
above. In LineageOS 14, there is a hack so it's not really required.
In LineageOS 15, the no one has been able to port the hack so this patch might
be needed.

This commit just re-encapsulates the metadata to use native_handle without passing
any fildes so that we won't need to close any. WRT this change, the
encoder may have to be tweaked to read this new encapsulation.
The commit f8c7c902abd1fdb53f3461672774fc68ae0cc5eb on this tree by
(coincidentally) djeman might not be complete in this case.
Honestly though, I don't know his/her motives so I can't say for sure.

NOTE:
On the bright side, video recording is now WORKING!!! 720p recording works too.
But also we've lost: autofocus, autoexposure, auto whitebalance, metering modes,
DTP and auto flash.

Change-Id: Ic1865500b2603c84f8645467ee08885aaa5c53d9
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>