libcamera: scx15: Apply SENSOR_IO_POWER_ONOFF on Sensor_PowerOn_Ex
This requires the patch named as
"libcamera: scx15: IOCTL extensions for some SPRD devices"

In Week 3 of the World-wide Quarantine due to the Coronavirus pandemic,

it seems that the SENSOR_IO_POWER_ONOFF was also called from the Sensor_PowerOn_Ex()
function not just the plain 'ol PowerOn function. Though this seems a useless
observation, this is a large part of the _cause_ why the autofocus on the
s5k4ec sensor fails.

Change-Id: I16e34acee379c254bdc1966fd9dd2833d8e36a0f
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>