hwcomposer: kanas: DONOTMERGE: Partially disable PROCESS_VIDEO_USE_GSP
GSP's blending YUV and RGB layers doesn't look right by a mile.
The "SPRDevs" (pun intended) are right for using the YUV420 format
for the output. RGB format just loses a lot of colors.

But, the problem of having some sort of long
vertical or horizontal "pixels" still exists. This problem
should have been barely noticable if the video looks real
(eg. real life shots) but its not on some CGI videos.

To be consistent with the previous versions of the rom and the stock, turn off
the GSP processing for YUV layers. PROCESS_VIDEO_USE_GSP also handles the
second side of DIRECT_DISPLAY_SINGLE_OSD_LAYER which we might want to have.
So, disabling PROCESS_VIDEO_USE_GSP is ng. Now, we have this patch.

Change-Id: I04e6eb0994ee32b1994b44b3ff3d723d0c67cd75
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>