libcamera: scx15: Add an alternate location to store the sensor init cache
"sensor init cache" is not actually stated anywhere in the code.
The Camera HAL does its identify over an array of possible sensor code
every time a camera is initialized or started.
It takes quite a long time to do that so the Camera HAL writes the
results of identifying all camera sensors to a file at
/data/misc/media/sensor.file

On the next identify, it can just load the file and skip the identify
process altogether, though each sensor will still be verified.

But, /data/misc is unreadable from the camera.provider service.
We can't just change the permission of data/misc since it is also used
by the Platform Code (e.g. AOSP) and changing it may bring some unintended
consequences.

However, we created the /data/local/media, initially for audio para files,
and coincedentally is accessible from camera.provider. So, let's use that.

Try the /data/misc one first and if it fail, go with /data/local/.

Change-Id: I958e5fd6feab9dcba046af4c1bc53d97a9b1620a
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>