camera: kanas: Implement 1280x960 record resolution; Correct AF window setting
The 1280x720 works well but may have a glaring issue, s5k4ecgx is a 4:3
camera sensor

1280x720 is 16:9. To get to that resolution, we remove, or trim ,
the excess of 4:3 usually on the top area. This may cause problems
like having an ironically smaller viewing area.

To alleviate that, 1280x960 preview resolution is implemented. It's
4:3 meaning you could get most of the sensor's input.

NOTE: Yep, 1080p should be possible to implement but the maximum FPS
we could get on kanas might be up to 20fps, 15fps average.

WARNING: An unknown cause somehow locks the AE when switching modes.
         Maybe due to a poorly canceled AF task.

This commit should have also bundled some fixes:
   AF window setting handles x,y coords as top-left corners rather than center
       This should fix the wrong focusing
   AF Window size can now be set by the camera app.
   Brightness have been toned done in 1280x720 (also in 1280x960)
   Additional capture/preview modes have been added
   [NOTE] Highest capture res has been brought down from 2592x1944 to 2560x1920
       This should prevent needless scaling in the sprd_dcam as
       2560x1920 is the max res the sensor can output

Change-Id: I61334571a17a91df3e9c85421f3613d855f675c4
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>