audio: kanas: Unify fast and "low power" pcm config
It was controlled using the DEEP_BUFFER flags which no one dares to use
since it breaks the audio playback altogether.

The nice thing about the DEEP_BUFFER or just simply using the low power pcm
config is that it helps decrease cpu usage when transferring frames
via DMA. It helps free up about 3-5% of cpu time (if that makes sense).

Though it increases the buffer and the latency by 3 times.
But it should be alleviated by changing the avail_min and setting the
start_threshold to a lower value. That speeds up the initial playback but the
latency problem remains when stopping it.

NOTE:
Longer buffers means less underruns? Nope, either by some weird delay
but even filling it in to the end of the buffer could still result in an
underrun... Broken Pipe. Unfortunately, this hardware have no way of
quickly recovering from that other than a reset which takes a second.
Unforgivable for music playbacks.

Change-Id: Ib8d469c7968038d320ebf5ce01d68cff459fe936
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>