hwcomposer: kanas: Don't use GSP when VSYNC is not enabled
TL;DR;
It causes screen tearing most often seen when displaying the
screen of animation on Android.

Full:
I haven't researched deeply enough why but it seems that the
DISPC overlay feature seems to be same to a double buffered display
with a non physically contiguous memory. This means that the
DISPC has to copy to a part of the frame buffer hopefully before
it also start displaying that current being-copied buffer.
Without VSync, there are no promises.

Change-Id: Ic73e1ffd405b41b5d3f4ac92d186d4889fe35cd2
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>