shim: engpc: Redirect /etc/audio_para to /vendor/etc/audio_para
Before the treble configuration, the file is actually on /etc/audio_para
which is a symlink to /system/etc/audio_para.

I don't know to what extent does it affect the running system when enpgc
cannot find this file. So, to be sure, intercept the open() calls and
fix any references to that file.

Change-Id: Ib5505566a75be8dbb0644c399ba3855e5b4ebe16
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>