shims: Add a new shim against the VSP library used in HW encoders
The newer (sc8830gX) encoder libraries (for h.264 AVC) can actually be used in
older hardware (for example, scx15), surprisingly. Porting the libstagefright
encoder code to the older platform is also straightforward.

But soon enough, one will see that the clocks aren't being set as intended.
The newer VSP lib expects the frequency table in the kernel to be sorted
in ascending order. The older platform's kernel has it in descending order.
One job of this shim is to translate the that IOCTL so that the newer VSP
lib's and the kernel driver's expectations are met.

In more recent VSP hardware, scene modes are supported. Of course, older
platforms do not support it but it can be faked so as to not raise
any error or suspension in the encoder.

Change-Id: Ie7aea56562f9550a777dcb0847f96cccb6b0adcb
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>