hwcomposer: kanas: Copy Overlay buffer to FB after the last frame; disable DIRECT_DISPLAY
This reverts the commit c9810d79a4bdedc6a0f1903b433947f9ce318c3c
(hwcomposer: kanas: Don't use GSP when VSYNC is not enabled)
since that's a bit too much to disable GSP on Vsync off.

We now try to copy the last applied buffer when display type has became unknown
then turn off the overlay feature after that. Really wished we could just simply put
off the GSP output to the FB but the physical memory address is not even possible
to retrieve from the kernel driver. I think it's possible, but it's tedious.

DIRECT_DISPLAY_SINGLE_OSD_LAYER is also now disabled for kanas as it brings forth a
problem from DISPC. It seemed that every frame DISPC tries to copy the overlay buffer
to the framebuffer(FB). The problem arises from the fact that the buffer is not persistent
at least until the HWC dies or crashes.

If ever the set overlay buffer (specifically it's address) is released in the host system
we can get some pretty funky display (or just simply garbage) until a new frame with a
new buffer arrives.

Change-Id: I69d86d7e79ab5ba2adccb83158381c193742485a
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>