hwcomposer: scx15: Allow disabling GSP on runtime
This behavior can be done by setting the property to 1
    debug.hwc.gsp.disable

By default, GSP is always enabled.

You could also just disable the video processing part by setting this to 1
    debug.hwc.gsp.video.disable

If the platform is kanas, the default value is 1, meaning it's disabled.
For others, it's allowed to process YUV/Video layers by default

Change-Id: Iad2d89beb468cfe3e508b70268af8c090bf92ef6
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>