camera: scx15: Fix TARGET_BOARD_CAMERA_DMA_COPY
This option is a simple optimization on cases where one has to copy the output buffer
of the sensor to the input of some native window. On why that case still happens
is still unknown to me. On some camera apps does that, some doesn't.
The default Snap camera app sometimes goes into the case where a copy is needed.

The optimization is to enforce a zero-copy operation when copying the two buffers
by setting up the DMA to do the job.

The problem is that the support is just broken:
   1. It does not allocate a physically contiguous memory.
   2. Fails to get the Physical Address of the said memory, if it were to allocate
      a physically contiguous block.
   2.1 It does gets a physical address but a wrong one. It's the magic address to the end of the rainbow.
   3. #2 does not cause an error and causes and out-of-bounds write...
   3.1 The out-of-bounds write will always be done with no errors,
       may or may not crash the whole system.

The kernel module, sprd_dma_copy, works just fine but the third problem is the most
worrying one yet.

This patch should properly fix those problems and even refactor the displayCopy()

Change-Id: I26e430583da3271d7d6ecaeac8507f6b71157c02
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>