gralloc: scx15: HIDL workaround to prevent ion orphaned allocations [WIP]
On HIDL after gralloc allocation, the buffer_handle is duplicated,
registered then freed... Yes, after allocating, even without using it,
it is freed. This breaks the traditional flow (read: life cycle) of
each buffer allocated via gralloc.

Fortunately, ion allows clients to close themselves even though the memory
allocated is still mmap'd. This allowed things to work as is though,
if you read the debugfs of ion heaps, you'll see most, if not all,
allocations are orphaned, and no one knows who's actually holding the buffer.

As I said earlier, it doesn't matter as the applications through HIDL will
still unregister the buffer in gralloc unmapping the memory and freeing
the buffer as intended.

This patch aims to fix these orphaned allocations by deferring the ion_free()
call -- the function that will close a client -- until the buffer is unregistered.
No need to wait until the last buffer unregister as it is not generally allowed
to close a client if it's not the client itself. Each duplicated handle
may have the same memory allocation but will have different clients opened
for them. So, we just free it until the creator unregisters it.

This patch is not yet complete.

Change-Id: If76a4a35acefbfe5eebf379a082b86753684ba52
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>