gralloc: scx15: Support formats HAL_PIXEL_FORMAT_YCbCr_420_P and HAL_PIXEL_FORMAT_YCbCr_422_I
Not much about HAL_PIXEL_FORMAT_YCbCr_422_I, but HAL_PIXEL_FORMAT_YCbCr_420_P
is needed to properly playback mpeg4 recordings.

If it's allocation, the pixel formats doesn't really matter just the size.
But the function does not specify the size, only the dimensions and format...
That's where this part of the allocation code goes into, as it's resposible
for determining the stride and size of the allocation,

HAL_PIXEL_FORMAT_YCbCr_420_P format uses the same size as it's siblings
in the 420 so not much changes. HAL_PIXEL_FORMAT_YCbCr_422_I on the other
hand consumes about 4 bits more per pixel compared to 420 so the size
has to be calculated a bit differently

Same as the previously defined formats, these two formats are 16 Bytes
aligned.

Change-Id: I7803f351d0d363e9d13f59a2bc37dab647e0c24c
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>