libcamera: scx15: Copy libmorpho_easy_hdr and libisp to /vendor/lib, for real
The previous commit with a simillar way uses an inappropiate variable
TARGET_COPY_OUT_VENDOR where as the actual variable is supposed to be
TARGET_OUT_VENDOR.

TARGET_OUT_VENDOR will contain an absolute file path to the directory
holding the /vendor partition. Though, it's /system/vendor for us
since we don't have a separate /vendor partition.

On the other hand, TARGET_COPY_OUT_VENDOR will just point to either
system/vendor or vendor. That's just it. It's a relative path but to where?
The top of the Android source tree.
If you're building the libcamera for scx15, Check the system/ folder where
the core folder from system/core project is located. There will be a
suspicious vendor folder there.

Change-Id: I565ab310c4ac94f022ced00e6977d620ec5687ef
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>