audio: scx15: Fix VBC firmware generation
The vbc_eq file at /data/local/media is autogenerated by the audio hal
everytime it starts and gets loaded by the hardware eventually.

In the stock rom, the /vendor/firmware/vbc_eq is actually a symlink to
/data/local/media and this regenerated everytime the audio hal is initialized.

In most of our custom roms, /vendor/firmware/vbc_eq is a plain ol file --
a precompiled one from the stock rom. Which should be fortunate to all of us.
Why?

Firmware generation in this source code is BROKEN. Loading this firmware will
effectively disable any sound, at least from the main speaker and headset.
Those using the NULL codec doesn't seem to be effected.

This patch erases the stub functions that should have been migrated to the
libvbeffects library. When not removed, these stub functions override
the full implementations at libvbeffects and prevents some
various parts of the firmware to be initialized
and configured properly.

Yep, even the function calls from libvbeffects gets stubbed.

Change-Id: I2781e507a28da4f98837606220cd6d9dfef64588
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>