gralloc: scx15: HIDL_DEFER_FREE FrameBuffer memory as well [WIP]
Framebuffer from gralloc aren't actually meant to be registered but with
the magic of the new-fangled HIDL, it's a impossibility made reality
along with directly freeing the handle after registering it.

Jokes aside, this patch will also defer freeing or releasing the
FB -- the gralloc's handle specifically. The problem with the current
HIDL_NO_FREE_FB is that you can only allocate up to 3 buffers from the FB
even after freeing all of them because the buffermask gralloc uses
never gets refreshed.

Combining with the logic of HIDL_DEFER_FREE, we could now restore the buffermask
allowing that case to happen.

Unfortunately, I have yet to see the SurfaceFlinger to unregister the FB.
It usually just allocates 3 buffers all at once and just cycles around without
unregistering or freeing any of them. In other terms, I cannot test this.

Change-Id: I6efb2c828d9af36610f3f22aeb74676d9a39b93f
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>