gralloc: scx15: Tweaks when only ION_OVERLAY as a carveout; Support for an extra heap
A lot of this are explain on the Android.mk and alloc_device.cpp so
I'll skim through.

There's an observation that vmalloc is slow and CMA is also slow and that
SurfaceFlinger often allocates and deallocates memory. By default,
without the TARGET_FORCE_HWC_CONTIG flag, SF buffers are vmalloc'd.
Then with TARGET_FORCE_HWC_CONTIG, SF buffers go through CMA.
An allocation on CMA is around 0-4 msecs, and  vmalloc 0-4 msec.
but it should be faster than CMA (if CMA would do some migration).
On a carveout memory, it should be under a couple microseconds.
vmalloc and CMA allocation success depends on whether the system can
still fullfil it. Carveout memory are outside of regular system usage
and should still be usable even under low memory conditions, provided
there are still some memory left in the carveout.

So... If we have SF get buffers from a craveout, not only we can have a
fast allocation regardless of memory pressure but also a physically
contiguous memory. That's the reason for the first one.

The extra heap support's raison d'etre is to have a fallback for failing
allocations. Like I said, carveouts are not affected by memory pressure
from the regular allocations in the system, let it be apps or something.
Specially for the camera, which takes at least 20MiB every single shot,
often fails when the system memory has few remaining contiguous blocks
and none fits the allocation requirement.

I may later on remove the extra heap support or improve on it after some
tests. To be honest, extra heap is dangerously incompatible with unpatched
kernels. So, -1 point for it?

Change-Id: I021541e61d4cee232b177ac9c31b607b07a7a757
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>