Revert "gralloc: scx15: HIDL workaround to prevent ion orphaned allocations [WIP]"
This reverts commit 5bff68ae0e909512632e51817afbffd881954e99.

The HIDL_DEFER_FREE workaround does not enable or fix any functionalities but
it introduces a particularly scary memory leak when the ION buffer is used like this
(I don't know if this is right.)
copy0: ALLOC->FREE->END
copy1:            ->REGISTER->...->UNREGISTER

Since we only call ion_free on unregister and the original copy0 didn't even register,
we effectively do not free the ion buffer and stays there untill the allocator dies.
Usually the allocator is SurfaceFlinger.

How to reproduce:
1. Open up any web browser app of your choice
2. Start noting the ION heap usage from this point.
3. Load any website with a video. Preferrably with autoplay and multiple videos in a page.
4. Wait until it finishes loading.
5. Close the web browser app.
6. Check the heap usage.
7. Repeat steps 1 to 6 and watch the usage climb up, refusing to go down.

For now, revert this patch.

Change-Id: Idf21575a938ecff20adfdc3d6819a2d77db66af3
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>