audio: kanas: Unified pcm config v1
From the binary dump of the audio.primary.sc8830.so, it is possible
to infer that Samsung did deviate  source a bit and this scx15
code provides a good common base.

The "struct pcm_config" config for normal playback (thru speakers or headset)
are divided into two. One for low latency and the other is for power saving (or
high latency).
When the DEEP_BUFFER flag is passed, this HAL driver will use the power saving
one other wise it's the other.
Samsung's binary blob does not have either thus does not support the feature.

The question that should have popped out would be what "struct pcm_config"
would they have used? Based on some quick pattern matching, there's only one
of it. What's funny is that what I've found closely matched the power saving
one or the high latency.

That configuration is written to pcm_config_alt and the buffer size is increased
from 87.0748 msec to 97.9592 msecs which is probably the longest and largest
buffer supported with SPRD IRAM.

Based on initial tests, cpu utilization when down to 0.5% from 0.9%,
comparing the previous patchset.

P.S. This is an experimental patch, akin to an [RFC], there are a couple of
properties that need to be set with setprop or through system.prop.
1. debug.audio.alternate_period
     When "1", this will use the pcm_config_alt above anything else, even
      if DEEP_BUFFER is on.
2. debug.audio.with_irq
     When "1", tinyalsa -- the underlying userspace alsa library -- will rely
     on the kernel (via poll()) to notify when the buffer needs to be
     filled. Usually, tinyalsa uses poll() with timeouts. This could be due to
     this HAL code being based from someone's audio HAL code or maybe because
     it satisfies deadlines much better than when the kernel is doing, thus
     preventing underruns.
     With IRQ's being boosted to SCHED_FIFO in the kernel, it should be possible
     to rely on the kernel. Doing so can impart a reduction on CPU utilization.

P.S.2 Both of these properties default to "1" on Kanas**(SM-G355H), "0" on other systems

Change-Id: I022757daa0c2df3bc946efaef8e47443a084cec9
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>