gralloc: scx15: Revert support for an extra heap; No fallbacks for OVERLAY Carveout
On the version 2 of the kernel patchset, the fallback mechanism is now
implemented on the kernel rather on the gralloc library. This helps
considering that some apps go through libion for allocations.

What was implemented for the extra heap is that it just adds some explicit fallback
which was also implemented on the kernel so there goes the dangerous flag
to the window.

Change-Id: I6733c90511f68dd15b1a0a048b3e55f2c394dd5b
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>