cameera: scx15: Fix compiler warnings
This is probably the largest codebase from SPRD and also "the"
nightmare when it comes to the  -Werror change from android 9.

It's difficult and wasteful to modify the source code just to iron
out ANY warnings so I started by ignoring these set of warnings:
unused-variable
unused-parameter
unused-function (only for the utest)

Then we set on fixing the rest of the warnings which makes sense.
An example of a warning that makes sense is the warning that appear
on a wrong format specifier in printf or even in ALOG functions, also
including too few or too many specifiers compared to the arguments passed.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>