libcamera: kanas: Disable flash algorithm after flash photography
The s_flash_state is updated every stream off and is the closest and the most
recent representation of whether the flash is off or on making it the worst
basis to check whether we still need to disable the flash algorithm in
the s5k4ecgx sensor. Not Disabling it will clamp the exposure time
usually causing the next photos on low light conditions to be dim
after the flash fires.

Before, s_flash_state was updated when we pretty much need it (a.k.a lazily
updated) so it suffices as an indicator that we need to disable it but
not today.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>