hwcomposer: scx15: Do not copy overlay buffer to FB at the last frame.
This partially reverts the commit titled
"hwcomposer: kanas: Copy Overlay buffer to FB after the last frame; disable DIRECT_DISPLAY"

I think I figured out how to make HWC2 via HWC2on1 adapter work but it opens
up to the possibilities that the gralloc and hwcomposer modules should able to
work with (or at least ignore) null buffer_handle_t objects. This feature
is a menace to that and doesn't actually (seem like it that it) does anything
noticable, let alone fix anything.

The DIRECT_DISPLAY bit is not reverted as it's just the limitation of the
method used. There's no way of telling the HWC that the layer buffer is already
gone, AFAIK.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I07c7ee3b5931a87ada8463392157262caff3a5b2