hwcomposer: scx15: Explicitly initialize all unused HWC1 callbacks to nullptr
This should (or would help) when ever the compiler refuses to zero initialize
and some module executes an otherwise unused callback just because it is not
actually nullptr.

Though this error only did pop out on LineageOS 16.0, and yes this is the branch
for LineageOS 15.1. I think this fix should also be in here. I assure you it is
not because I am testing AOSP 9 (LineageOS 16.0) modules on LineageOS 15.1.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: Ia85b8e67efb7d20a338c9c691cd841e93f2b4494