hwcomposer: Scan for display attributes until HWC_DISPLAY_NO_ATTRIBUTE is found
The default one is fixed to the point that it only assumes 5 attributes to scan.
HWC2on1 adapter seems to use this as a test to check whether the HWC supports
color transforms(?) and checks if the HWC fails when it tries to get attribute
including the color transform one.

Since the list is fixed, the color transform attribute is ignored and this test
should pass whilst returning 0 transforms.

This patch converts the NUM_DISPLAY_ATTRIBUTES to LIMIT_DISPLAY_ATTRIBUTES
which sets an upper limit to the number of attributes to be scanned.
The main breaking condition now is to check whether HWC_DISPLAY_NO_ATTRIBUTE is
found or we have reached LIMIT_DISPLAY_ATTRIBUTES.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I1b1d693a31b739aa71aec8e59a46cd4ba9b872dd