gralloc: scx15: A HIDL workaround to make fb_post() work between instances
In my testing, it is possible to have an instance (a HIDL one) and another
private instance (in HWC) the HIDL based instance creates a FB buffer_handle_t
passed to SurfaceFlinger. This will then be passed to the HWC with its own
gralloc instance. The HWC would most often pass this to fb_post which will
call an IOCTL to do pageflipping.

Unfortunately, the code that calculates the offsets would be incorrect because
the bases addresses used are different between the two instances.

This can be solved in two ways, store the base used in each buffer handle and
rewrite the HWC code call the gralloc module via HIDL.

The first option, has a problem since we cannot add nor remove any INT or FD in
the buffer_handle_t or else libMali.so will consider the buffer as invalid. But,
we could use an unused INT since not all INTS are used...

The second option, is the better solution... in paper. This is because there is
no HIDL way of accessing GRALLOC_HARDWARE_FB0 as far as I know so this solution
is not implementable.

With the workaround active, offset calculation would not be performed using the
stored base address from the buffer handle themselves.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: I47833cf549077b1247abb24b1c157f9a1f4cfd86