audio: scx15: Make audio_vbc_eq target recreate the firmware/vbc_eq link always
The problem I have with the audio_vbc_eq is that it sometimes does not create
the /vendor/firmware/vbc_eq link that the audio kernel driver loads.
This firmware is autogenerated by libvbeffect and is saved at /data/local/media.
Of course, saving to /vendor/firmware/ is not possible due to the /system/ being
mounted as read only.

The fake package that was written originally only works once when audio_vbc_eq
is built (or FAKED). After that, any succeeding builds will not regenerate
the vbc_eq even if the symlink is removed from the temporary /system/ directory
in the out/target/product/<product>/system. You have to search and destroy
the audio_vbc_eq immediates directory from out/target/product/<product>/obj
directory.

Enough talk.

This patch rewrites the Makefile to use the newer LOCAL_INSTALLED_MODULE_STEM
directive (which overrides the name of the created file) and sets it directly at
the expected path /system/vendor/firmware (/vendor is linked to /system/vendor
anyways).

Semantically, this rewrite first creates a blank file at
/system/vendor/firmware/vbc_eq. Then the post install commands overwrites the
blank file to a broken symlink. It is 'broken' in the host system but should be
a valid symlink in the device later on. We need to create a blank file first
because broken symlinks often breaks cp as it tries to follow the reference link
rather than copying the actual link.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>
Change-Id: Ib5199e1d91fb6a8087677f08bdff2e8396a0bf55